Don't redistribute link-local addresses.
We should deal with link-local addresses though, but it's a special
case which needs to lookup at the interface index and so on.  Not
specially easy to use, so I prefer discard it for now.  But it would
be great to keep the connection to a router, even after a manipulation
error which have killed the router's addresses (for example).