Map flows by ID.
The server will have the same flow ID than the client, for each
flow.  This way, if the client use an explicit flow (src, dst)
with a "system" flow (_, dst), and if the kernel decide to use
"src" as the source, then the two flows will now be consider
different by the client.